next-market-ddb-login/requirements.txt
requests
PyJWT
aws-xray-sdk